Crypto Clicker : fusion étendue + bonus éléments permanents
- PeriodicCollectionStore : compteur cumulatif total (gacha+fusion) et
  compteur fusion-only ; addCopyFromFusion() incrémente les deux
- ElementBonusEngine : flat via fusion uniquement (sauf H/He via gacha) ;
  progression x10 par palier C+10, Ne+10, O+100, Fe+100 ; SUPERNOVA
  Z15-28 +1000 APC/APS depuis neutron capture ; % calculés sur copies
  gacha uniquement pour les éléments de la chaîne
- FusionRecipe : WATER renommé SULFUR_BURNING (id conservé pour saves) ;
  FusionOutput.RandomRarity ajouté ; NEUTRON_CAPTURE palier 7
  (3 Fe → élément Supernova aléatoire, 12%)
- FusionActivity : résolution de l'élément aléatoire au moment du succès,
  appel addCopyFromFusion(), affichage du résultat pour RandomRarity
- Strings EN + FR pour NEUTRON_CAPTURE

+++ app/src/main/java/com/Atom2Universe/app/clickerstats/ClickerStatsActivity.kt
@@ -200,7 +200,7 @@ class ClickerStatsActivity : ThemedActivity() {
     }
 
 
-    private fun formatElementBonus(flat: Int, mult: Double): String {
+    private fun formatElementBonus(flat: Long, mult: Double): String {
         val parts = mutableListOf<String>()
         if (flat > 0) parts.add("+$flat flat")
         if (mult > 0.0) parts.add("+${"%.2f".format(mult * 100)}%")
